requirements.txt
Flask==3.0.0
Flask_RESTful==0.3.10
mysql_connector_repackaged==0.3.1
gunicorn